+++ src/prelude.rs
@@ -32,8 +32,8 @@ pub use crate::{
         JustifyContent, JustifyItems, JustifySelf, LengthPercentage, LengthPercentageAuto, Position, Style,
     },
     style_helpers::{
-        auto, fit_content, flex, max_content, min_content, minmax, percent, points, zero, FromFlex, FromPercent,
-        FromPoints, TaffyAuto, TaffyFitContent, TaffyMaxContent, TaffyMinContent, TaffyZero,
+        auto, evenly_sized_tracks, fit_content, flex, fr, max_content, min_content, minmax, percent, points, zero,
+        FromFlex, FromPercent, FromPoints, TaffyAuto, TaffyFitContent, TaffyMaxContent, TaffyMinContent, TaffyZero,
     },
     tree::LayoutTree,
 };